# python/rock/projeto03/swagger.yaml
openapi: 3.0.0
info:
  title: Projeto03 - Tasks API
  version: "1.0.0"
  description: API simples de tarefas (tasks) do projeto03
servers:
  - url: http://localhost:5001
paths:
  /tasks:
    post:
      summary: Criar nova tarefa
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Tarefa criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    get:
      summary: Listar tarefas
      tags:
        - tasks
      responses:
        '200':
          description: Lista de tarefas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'

  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: ID da tarefa
    get:
      summary: Obter tarefa por ID
      tags:
        - tasks
      responses:
        '200':
          description: Tarefa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Tarefa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    put:
      summary: Atualizar tarefa
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Tarefa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Tarefa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      summary: Remover tarefa
      tags:
        - tasks
      responses:
        '200':
          description: Tarefa removida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Tarefa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        done:
          type: boolean
      required:
        - id
        - title

    TaskCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        done:
          type: boolean
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        done:
          type: boolean

    TaskList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total_tasks:
          type: integer

    Message:
      type: object
      properties:
        message:
          type: string